dev-ps-toilet-app: refactor code
